+++ game/commander/tasks/primitive/antishipping.py
@@ -22,5 +22,6 @@ class PlanAntiShipping(PackagePlanningTask[CargoShip]):
         state.enemy_shipping.remove(self.target)
 
     def propose_flights(self) -> None:
-        self.propose_flight(FlightType.ANTISHIP, randint(2, 4))
+        size = self.get_flight_size()
+        self.propose_flight(FlightType.ANTISHIP, size)
         self.propose_common_escorts()