+++ game/commander/tasks/primitive/cas.py
@@ -30,5 +30,6 @@ class PlanCas(PackagePlanningTask[FrontLine]):
         state.vulnerable_front_lines.remove(self.target)
 
     def propose_flights(self) -> None:
-        self.propose_flight(FlightType.CAS, randint(2, 4))
+        size = self.get_flight_size()
+        self.propose_flight(FlightType.CAS, size)
         self.propose_flight(FlightType.TARCAP, 2)